Cleanup includes
Might improve compile times slightly.

Also adds a CMake target for the matlab extension, making it easier to check for compilation errors there.
