+++ include/amici/model_state.h
@@ -293,6 +293,24 @@ struct ModelStateDerived {
 };
 
 
+/**
+ * @brief implements an exchange format to store and transfer the state of a simulation at a
+ * specific timepoint.
+ */
+struct SimulationState{
+    /** timepoint */
+    realtype t;
+    /** state variables */
+    AmiVector x;
+    /** state variables */
+    AmiVector dx;
+    /** state variable sensitivity */
+    AmiVectorArray sx;
+    /** state of the model that was used for simulation */
+    ModelState state;
+};
+
+
 } // namespace amici
 
 #endif // AMICI_MODEL_STATE_H
