+++ include/amici/steadystateproblem.h
@@ -1,11 +1,10 @@
 #ifndef AMICI_STEADYSTATEPROBLEM_H
 #define AMICI_STEADYSTATEPROBLEM_H
 
-#include "amici/defines.h"
-#include "amici/forwardproblem.h"
-#include "amici/solver_cvodes.h"
-#include "amici/vector.h"
+#include <amici/defines.h>
+#include <amici/vector.h>
 #include <amici/newton_solver.h>
+#include <amici/model_state.h>
 
 #include <nvector/nvector_serial.h>
 
@@ -339,27 +338,27 @@ class SteadystateProblem {
      * dampening (false)
      */
     bool updateDampingFactor(bool step_successful);
-    
+
     /**
      * @brief Updates member variables to indicate that state_.x has been updated and xdot_, delta_, etc.
      * need to be recomputed.
      */
     void flagUpdatedState();
-    
+
     /**
      * @brief Retrieves simulation sensitivities from the provided solver and sets the corresponding flag
      * to indicate they are up to date
      * @param solver simulation solver instance
      */
     void updateSensiSimulation(const Solver &solver);
-    
+
     /**
      * @brief Computes the right hand side for the current state_.x and sets the corresponding flag to
      * indicate xdot_ is up to date.
      * @param model model instance
      */
     void updateRightHandSide(Model &model);
-    
+
     /**
      * @brief Computes the newton step for the current state_.x and sets the corresponding flag to
      * indicate delta_ is up to date.
@@ -449,14 +448,14 @@ class SteadystateProblem {
     bool newton_step_conv_ {false};
     /** whether sensitivities should be checked for convergence to steadystate */
     bool check_sensi_conv_ {true};
-    
+
     /** flag indicating whether xdot_ has been computed for the current state */
     bool xdot_updated_ {false};
     /** flag indicating whether delta_ has been computed for the current state */
     bool delta_updated_ {false};
     /** flag indicating whether simulation sensitivities have been retrieved for the current state */
     bool sensis_updated_ {false};
-    
+
 };
 
 } // namespace amici
