+++ src/interface_matlab.cpp
@@ -8,9 +8,11 @@
 
 #include "amici/interface_matlab.h"
 
+#include "amici/amici.h"
 #include "amici/model.h"
 #include "amici/exception.h"
 #include "amici/edata.h"
+#include "amici/solver.h"
 #include "amici/returndata_matlab.h"
 
 #include <assert.h>
@@ -120,7 +122,7 @@ std::unique_ptr<ExpData> expDataFromMatlabCall(const mxArray *prhs[],
     if (!mxGetPr(prhs[RHS_DATA]))
         return nullptr;
 
-    auto edata = std::unique_ptr<ExpData>(new ExpData(model));
+    auto edata = std::make_unique<ExpData>(model);
 
     // Y
     if (mxArray *dataY = mxGetProperty(prhs[RHS_DATA], 0, "Y")) {
